Add: debug statements to check for latest post file

+++ src/slack_notification.py
@@ -4,8 +4,7 @@ import os
 
 # Define the RSS feed URL and Slack webhook URL
 rss_url = "https://www.bunnieabc.com/index.xml"
-slack_url = os.environ['SLACK_WEBHOOK_URL'] = "https://hooks.slack.com/services/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"
-
+slack_url = os.environ['SLACK_WEBHOOK_URL']
 
 # Define the notification message format
 message_format = "New post published on {blog_title}:\n<{post_url}|{post_title}>"
@@ -15,25 +14,27 @@ feed = feedparser.parse(rss_url)
 latest_post = feed.entries[0]
 
 # Check if the latest post is different from the previous one
-with open('latest_post.txt', 'r+') as f:
-    previous_post_link = f.read()
-    if latest_post.link != previous_post_link:
-        # Construct the notification message
-        message = message_format.format(blog_title=feed.feed.title, post_title=latest_post.title, post_url=latest_post.link)
+if os.path.isfile('src/latest_post.txt'):
+    with open('src/latest_post.txt', 'r+') as f:
+        previous_post_link = f.read()
+        if latest_post.link != previous_post_link:
+            # Construct the notification message
+            message = message_format.format(blog_title=feed.feed.title, post_title=latest_post.title, post_url=latest_post.link)
+            print("New post found:\n", message)
 
-        # Send the notification to the Slack webhook URL
-        response = requests.post(slack_url, json={"text": message})
+            # Send the notification to the Slack webhook URL
+            response = requests.post(slack_url, json={"text": message})
 
-        # Check the response status code to ensure the notification was sent successfully
-        if response.status_code == 200:
-            print("Notification sent successfully!")
-            # Write the link of the latest post to the file
-            f.seek(0)
-            f.write(latest_post.link)
-            f.truncate()
+            # Check the response status code to ensure the notification was sent successfully
+            if response.status_code == 200:
+                print("Notification sent successfully!")
+                # Write the link of the latest post to the file
+                f.seek(0)
+                f.write(latest_post.link)
+                f.truncate()
+            else:
+                print("Error sending notification:", response.status_code)
         else:
-            print("Error sending notification:", response.status_code)
-    else:
-        print("No new post found.")
-
-#DONE        
+            print("No new post found.")
+else:
+    print("latest_post.txt file not found!")
